Census_tract_API_Data_extraction
This program stores the Census tracts data from https://api.census.gov/data/2020/acs/acs5?get=NAME&for=tract:*&in=state:01 for all the states in the nation by iterating the state number and stores the data in a csv file. Pandas data frames are used to clean the csv file.